requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
autoflake==2.3.1
black==24.10.0
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
flake8==7.1.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
isort==5.13.2
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
pyasn1==0.6.1
pycodestyle==2.12.1
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyflakes==3.2.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.36
starlette==0.41.3
text-unidecode==1.3
typing_extensions==4.12.2
Unidecode==1.3.8
uvicorn==0.34.0
WTForms==3.1.2
